# docs/swagger.yaml
definitions:
...
  handlers.CreateRoomRequest:
    properties:
      isPrivate:
        example: false
        type: boolean
      name:
        example: General Chat
        type: string
      slug:
        example: general
        type: string
    required:
    - name
    - slug
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
    type: object
...
  handlers.MessageResponse:
    properties:
      createdAt:
        example: "2024-01-15T10:30:00Z"
        type: string
      id:
        example: 1
        type: integer
      imageUrl:
        example: https://example.com/image.jpg
        type: string
      reactions:
        additionalProperties:
          items:
            type: integer
          type: array
        type: object
      roomId:
        example: 1
        type: integer
      text:
        example: Hello everyone!
        type: string
      type:
        example: text
        type: string
      userId:
        example: 1
        type: integer
    type: object
...
  handlers.RoomResponse:
    properties:
      id:
        example: 1
        type: integer
      isPrivate:
        example: false
        type: boolean
      name:
        example: General Chat
        type: string
      ownerId:
        example: 1
        type: integer
      slug:
        example: general
        type: string
      unread:
        example: 5
        type: integer
    type: object
  handlers.SendMessageRequest:
    properties:
      imageUrl:
        example: https://example.com/image.jpg
        type: string
      text:
        example: Hello everyone!
        type: string
      type:
        example: text
        type: string
    type: object
  handlers.SuccessResponse:
    properties:
      ok:
        example: true
        type: boolean
    type: object
...
  handlers.UserResponse:
    properties:
      avatarUrl:
        example: https://example.com/avatar.jpg
        type: string
      id:
        example: 1
        type: integer
      lastSeen:
        example: "2024-01-15T10:30:00Z"
        type: string
      login:
        example: john_doe
        type: string
      name:
        example: John Doe
        type: string
      online:
        example: true
        type: boolean
    type: object
...
info:
  contact: {}
paths:
...
  /rooms:
    get:
      description: Возвращает список всех доступных комнат
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.RoomResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить список комнат
      tags:
      - rooms
    post:
      consumes:
      - application/json
      description: Создает новую комнату для чата
      parameters:
      - description: Данные комнаты
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.RoomResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать комнату
      tags:
      - rooms
  /rooms/{id}/join:
    post:
      description: Присоединяет текущего пользователя к указанной комнате
      parameters:
      - description: ID комнаты
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Присоединиться к комнате
      tags:
      - rooms
  /rooms/{id}/leave:
    post:
      description: Удаляет текущего пользователя из указанной комнаты
      parameters:
      - description: ID комнаты
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Покинуть комнату
      tags:
      - rooms
  /rooms/{id}/messages:
    post:
      consumes:
      - application/json
      description: Отправляет новое сообщение в указанную комнату
      parameters:
      - description: ID комнаты
        in: path
        name: id
        required: true
        type: string
      - description: Текст сообщения
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/handlers.SendMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Отправить сообщение
      tags:
      - messages
...
  /rooms/{id}/read:
    post:
      parameters:
      - description: ID комнаты
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Отметить комнату как прочитанную
      tags:
      - rooms
  /rooms/{id}/users:
    get:
      description: Возвращает список всех пользователей в указанной комнате
      parameters:
      - description: ID комнаты
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.UserResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Список пользователей в комнате
      tags:
      - rooms
...
swagger: "2.0"